Change inf_cl location
