+++ ops/src/paddlenlp_kernel/triton/inf_cl/__init__.py
@@ -13,5 +13,4 @@
 # limitations under the License.
 
 from .flash import cal_flash_loss
-from .inf_cl_loss import *
 from .ring import cal_inf_loss, cal_ring_loss
